Addressed new comments
